+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Switch to `opendal` for async benchmark
+
 ## [0.4.2] - 2024-05-16
 
 ### Changed